APP/requirements.txt
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dash==2.17.0
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
fiona==1.9.6
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
fonttools==4.52.1
geopandas==0.14.4
greenlet==3.0.3
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
plotly==5.22.0
psycopg2-binary==2.9.9
pyparsing==3.1.2
pyproj==3.6.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.2
retrying==1.3.4
shapely==2.0.4
six==1.16.0
SQLAlchemy==2.0.30
tenacity==8.3.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
zipp==3.18.2
